Allows for migration of a single file
The default is now to only migrate files explicitly passed as
entrypoints. Passing the `-r` flag will migrate all dependencies
recursively.

We still need to look at all dependencies even if we're not migrating
them. To make this cleaner, the migration now has two passes.

First, we build the dependency graph by parsing the StylesheetApi
(consisting of the variables, functions, mixins, and imports) of
every sheet in the graph.

We then run the actual migration of the entrypoints (and, if desired,
their dependencies).