Support migrating built-in fns to use modules
When built-in functions are used within a source file, the migrator will
now add an import for the corresponding built-in module and add a
namespace to the function call.

As part of this change, the migrator now finishes the file it is on
before starting the migration of any dependencies. This eliminates the
need to store in-progress migration data for more than one file at a
time.