Various refactors based on review on #4.
Highlights:
- The migrator now extends RecursiveStatementVisitor, eliminating
the need for BaseVisitor
- Migration state for each stylesheet is now stored in
StylesheetMigration, which keeps track of member scope, namespaces,
and patches.
- The migrator now always runs on all dependencies, with the
command-line options now only used to determine which migrations are
written to disk. This simplifies the migrator logic.
- Since dependencies are migrated at the time the import rule is
migrated, the global members of the dependency are forwarded onto the
StylesheetMigration, which allows the namespacing of implicit
dependencies to avoid the complicated logic of examining the entire
dependency graph each time.
- The tests now use test_descriptor to avoid the need for mocks.