Add basic support for namespacing get-function
Resolves #11.

If the argument passed to get-function is a string literal, we'll add a
namespace if necessary.

For all other get-function calls, we'll emit a warning saying that a
namespace may be required. We might eventually be able to detect some of
these with some more sophisticated logic here, but for now we'll just
handle the basics.