Initial implementation of a division migrator
This adds a new migrator that migrates the `/` operator to the division
function.

By default, this migrator migrates only cases where the `/` is known to
be division. In cases the `/` could be division, but whether it is
division is not statically known, a warning is emitted instead. The
`--aggressive` flag can be used to migrate cases like `3 / $x + 1`, `3 /
$x - 1`, and `fn() / 3` by assuming that `+` and `-` always operate on
numbers and that function calls in an expression with `/` always return
numbers.