Initial implementation of a division migrator (#17)
* Generalizes to support multiple migrators

There's now a `Migrator` class that additional migrators can extend.

Most migrators will want to extend `MigratorBase`, which handles the
tree traversal. For now, `MigratorBase` only operates on a single
stylesheet, with the dependency migration implemented only by
`ModuleMigrator`.

* Generalizes more of the migrator

--migrate-deps is now a global flag, so additional migrators should not
need to re-implement dependency migration unless they need additional
special logic like the module migrator needs.

The generic migration classes are now split into Migrator and
MigrationVisitor. Each Migrator is a Command and should contain state
for the entire migration run. Most migrators will also want to create a
private subclass of MigrationVisitor that they call from
Migrator.migrateFile. A new instance of the MigrationVisitor will be
created for each stylesheet being migrated, so it contains
per-stylesheet state, similar to what StylesheetMigration used to be for
the module migrator.

* Split migrator tests into separate test files

* Minor formatting fixes

* Iterate in response to review.

The migrator now uses URLs instead of paths, and loads files using the
FilesystemImporter.

`Migrator.migrated` has been removed in favor of having
`Migrator.migrateFile` and `MigrationVisitor.run` return it.

* Reuse single MigrationVisitor

A single MigrationVisitor is now reused instead of constructing a new
instance for each dependency.

This also removes support for configurable variables, since the previous
implementation of this was both hard to implement with a single
MigrationVisitor and incorrect when the configurable variables were
declared in an indirect dependency.

A subsequent PR will add this back with a new implementation that fixes
the bug with indirect dependencies.

* Initial implementation of a division migrator

This adds a new migrator that migrates the `/` operator to the division
function.

By default, this migrator migrates only cases where the `/` is known to
be division. In cases the `/` could be division, but whether it is
division is not statically known, a warning is emitted instead. The
`--aggressive` flag can be used to migrate cases like `3 / $x + 1`, `3 /
$x - 1`, and `fn() / 3` by assuming that `+` and `-` always operate on
numbers and that function calls in an expression with `/` always return
numbers.

* Changes in response to review

Replaced --aggressive with --pessimistic, with the previous behavior for
aggressive now the default. Also made other minor changes in response to
review.

* Remove unnecessary parens around args to division

* Refactor `/` patching into _visitSlashOperation

* Convert non-division `/` operations to slash-list

* Address remaining comments on PR