Put Arguments in variables scope instead of null.
Putting null in the variables map for CallableDeclaration arguments
was causing errors, so we expand the type of it to allow any SassNode.

This makes the type of Scope.variables less specific than we'd like it
to be, but we don't actually do anything that assumes it contains
VariableDeclarations anywhere else.