Run Dart VM tests with snapshot
Similar to how the Node tests require pub run grinder js to be run, the
Dart VM tests now require pub run grinder app-snapshot to be run,
drastically decreasing the amount of time it takes to run all of the
tests.

The new `snapshot` tag is necessary for these tests now since filtering
with `-t node` will still cause the check for the snapshot's existence
to be run, but `-x snapshot` won't.