Fix bug with configurable variables
Previously, if References encountered the same !default variable
declaration multiple times (e.g. through importing the same file
multiple times), it would consider it to have configured itself.

References would also consider a variable "configured" even if it was
overridden in the same stylesheet as the !default declaration.

This fixes both of these so that declarations are only added to
defaultVariableDeclarations when their source URL differs from that of
the existing one.