Support entrypoints with migrated dependencies
The module migrator should now work on stylesheets whose dependencies
have already been migrated to the module system. _ReferencesVisitor will
visit all dependencies (including pre-migrated ones) to determine where
all members are defined, but the main migrator will not visit
stylesheets depended on via `@use` or `@forward`.

To track forwarded members, this adds the `Forwarded` interface, which
wraps a `VariableDeclaration`, `MixinRule`, or `FunctionRule`, but
reports its span as that of the `@forward` rule to ensure that
references to forwarded members are namespaced appropriately.