Handle references that appear before declaration
It's possible for references within functions and mixins to appear
before the declaration, so we need to track these separately.

In References, we now track references that couldn't be resolved at the
time they were visited and we then check to see if they've been resolved
when by a subsequent declaration at the time we close the scope that
contains the declaration.

Similar, in the main visitor, we now check to see if globalDeclarations
need to be renamed before we actually visit any stylesheets. We then
patch both declarations and references based on _renamedMembers.