Track references to migrate pseudoprivate members (#76)
* Adds first-pass to module migrator to track refs

Before the main MigrationVisitor runs, the module migrator now generates
a References object which maps member declarations to references.

* Remove dash from pseudoprivate members

This significantly changes the way members are renamed. Most tests are
still passing, but need to make changes to fix issues with reassigning
to a variable.

* Fix remaining issues with new code for renaming

This changes _scope to use the member's original names instead of what
their declarations were renamed to. Renamed members are now tracked in
_renamedMembers.

* Fix comments

* Use ImportCache; misc code review changes

* Misc changes from code review

* Switch back to renaming references when visited

* Resolve relative URLs with ImportCache

* Remove scope tracking from main migrator pass

Scope tracking is now only done in the initial pass (that generates
References). To make this work, References now also tracks global
declarations and !default variable declarations, and the main migrator
separately tracks unreferencable members (from nested imports) and
variables that are reassigned in a different module.

The unreferencable-member-tracking part of Scope has been refactored
into a separate UnreferencableMember class used by the main migrator.

* Fix importing entrypoints

The previous behavior would have allowed dependencies to be resolved
relative to the directory the executable is run from.

* Fix bugs in finding name spans

* Visit function arguments in References

Forgot to call super.visitFunctionExpression

* Normalize function names in references

Unlike variables and mixins, FunctionExpression's `name` property
is not normalized, so underscores are left as is.

* Fix bug with configurable variables

Previously, if References encountered the same !default variable
declaration multiple times (e.g. through importing the same file
multiple times), it would consider it to have configured itself.

References would also consider a variable "configured" even if it was
overridden in the same stylesheet as the !default declaration.

This fixes both of these so that declarations are only added to
defaultVariableDeclarations when their source URL differs from that of
the existing one.

* Handle references that appear before declaration

It's possible for references within functions and mixins to appear
before the declaration, so we need to track these separately.

In References, we now track references that couldn't be resolved at the
time they were visited and we then check to see if they've been resolved
when by a subsequent declaration at the time we close the scope that
contains the declaration.

Similar, in the main visitor, we now check to see if globalDeclarations
need to be renamed before we actually visit any stylesheets. We then
patch both declarations and references based on _renamedMembers.

* Misc changes for code review