Follow Unix-y error conventions
This sets a non-zero exit code when an error is encountered, and
prints error messages and warnings to stderr.

This also expands the test HRX semantics: stdout is still specified in
log.txt, but stderr is specified in warning.txt (if the migration is
expected to succeed) or error.txt (if it's not).

Closes #14