Error if renaming member in unmigrated file
When running the module migrator without `--migrate-deps`, it's
possible that it may generate invalid member references if it renames
a member from a dependency that is not being migrated.

This adds an error when this happens that tells the user to re-run the
migrator with `--migrate-deps` or add the dependency as an entrypoint.