Add parens around negated variable when namespaced (#115)
Fixes #114.

In fixing this, I also found a fixed a bug when migrating removed color
functions, as a patch to namespace the argument would be applied before
the patch adding the new argument name. This was fixed by ensuring the
arguments of a function expression are visited after the function itself
is patched.