Fix issue with ruleURL
We were previously setting ImportSource.ruleURL to the URL of the
`@forward` rule for members forwarded through import-only files, but
this breaks path-based namespacing, since that rule's URL is relative to
the import-only file, rather than the file making the import.

ruleURL is now set to the `@import` rule's URL for forwards of the
regular file for the import-only file. For other forwards, ruleURL is
left null until it is set based on _absoluteUrlToDependency in
_determineNamespaces.