Refactor ModuleMember
This shrinks the API surface and clarifies exactly which URLs come
from where. It also creates a type distinction for import-only
members, so that their attributes can only be accessed when they're
known to be valid.

This in turn fixes a few bugs with modules that were forwarded through
multiple layers of import-only files.