Adds the namespace migrator
This allows you to change namespaces for `@use` rules by matching
regular expressions on the existing namespace or on the rule URL.

The migrator also has special handling for `@use` rules without member
references which can be configured to run only when they conflict with a
referenced `@use` rule (default), for all unreferenced rules, or for
none of them.

The renaming itself is specified through a small DSL. The basics of this
are documented in the changelog and the help text, but I'll add proper
documentation to the website before merging this and releasing 1.3.0.

The tests under `test/migrators/namespace` focus on conflict resolution
and the actual patching of source files, while the DSL itself is mostly
tested in the unit tests in `test/renamer_test.dart`.