Fix a crash with orphan import-only files
Previously we just used the rule URL from the orphan import-only's last
`@forward` rule directly, but that only works when that URL is relative
to a load path or the orphan is in the same directory as the file
depending on it.

We now build a new dependency URL after resolving the old rule URL into
its canonical form.